+++ ski-jump-manager/src/components/ButtonsBox.jsx
@@ -3,6 +3,7 @@ import '../app.scss';
 import playGamePic from '../assets/images/playGamePic.png';
 import tutorialPic from '../assets/images/tutorialPic.png';
 import leavePic from '../assets/images/leavePic.png';
+import { Link } from "react-router-dom";
 const { ipcRenderer } = window.require('electron')
 
 export default function ButtonsBox() {
@@ -12,7 +13,7 @@ export default function ButtonsBox() {
 
     return (
         <div className="buttons-box">
-            <button> <img src = {playGamePic} /> Rozpocznij grę </button>
+            <Link to = '/chooseCountry'> <button> <img src = {playGamePic} /> Rozpocznij grę </button> </Link>
             <button> <img src = {tutorialPic} /> Samouczek </button>
             <button onClick={closeApp}> <img src = {leavePic} /> Zamknij grę </button>
         </div>