+++ ski-jump-manager/src/main.jsx
@@ -2,11 +2,16 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import {createBrowserRouter, RouterProvider} from 'react-router-dom'
 import StartPage from './pages/StartPage.jsx'
+import ChooseCountryPage from './pages/ChooseCountryPage.jsx'
 
 const router = createBrowserRouter([
   {
     path: '/',
     element: <StartPage />
+  },
+  {
+    path: '/chooseCountry',
+    element: <ChooseCountryPage />
   }
 ])