+++ CHANGELOG.md
@@ -1,10 +1,22 @@
+## [0.0.3] - 2024-02-25
+
+### Added
+ - Added choose country page.  
+ - Media queries for choose country page is done (min width: 1200px).
+ - Added function which generate a savegame folder and savegame.json (fs).
+
+### Corrected
+ - Sass were optymalized.
+
+## ________________________________________________________________________________________________
+
 ## [0.0.2] - 2024-02-24
 
 ### Added
  - Add start page.  
  - Media queries for start page is done (min width: 1200px).
 
-## ________________________
+## ________________________________________________________________________________________________
 
 ## [0.0.1] - 2024-02-23
  - Initial project. Installed react, electron, vite, sass. checked action of ipcMain (correct).